% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coveredMarkerProtein.R
\name{calculateCoveredProtein}
\alias{calculateCoveredProtein}
\title{Evaluate marker protein coverage}
\usage{
calculateCoveredProtein(proteinIDs, markerproteins)
}
\arguments{
\item{proteinIDs}{character; gene symbol id}

\item{markerproteins}{character; 3365 proteins gene symbol ids}
}
\value{
covered.proteins
}
\description{
Given the proteomics data, number of overlapped marker
proteins is calculated.
Bar plot for each compartment is plotted.
}
\examples{
{

df <- loadData(SubCellBarCode::hcc827Ctrl)

c.prots <- calculateCoveredProtein(rownames(df), markerProteins[,1])
}
}
