% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calRowMean.R
\name{calRowMean}
\alias{calRowMean}
\title{Compute the means of replicates}
\usage{
calRowMean(d.df)
}
\arguments{
\item{d.df}{data.frame; A data frame of 10 fraction profiles
consisting of replicate A and B.}
}
\value{
r.df
}
\description{
Duplicated franctions A and B are
summarized by taking their mean for each protein.
After taking the mean, the data log2 transformed.
Further, the 5 main fractions are used to check correlation
between input datas.
}
\examples{
{

r.df <- calRowMean(SubCellBarCode::hcc827Ctrl)

}
}
