% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc827Ctrl.R
\docType{data}
\name{hcc827Ctrl}
\alias{hcc827Ctrl}
\title{HCC827 Control Cell Line}
\format{A data frame where 10480 protein gene-centric ids and
5 replicated subcellular fractions.}
\usage{
hcc827Ctrl
}
\description{
Subcellular fractionated cell line.
}
\examples{
{
head(hcc827Ctrl)
}
}
\references{
Orre et al. 2019 Cell 73, 1-17
}
\keyword{datasets}
