% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc827CtrlPSMCount.R
\docType{data}
\name{hcc827CtrlPSMCount}
\alias{hcc827CtrlPSMCount}
\title{Minimum PSM Count in HCC827Ctrl Cell Line.}
\format{
A data frame where 10480 protein gene-centric ids minimum PSM count.
}
\usage{
hcc827CtrlPSMCount
}
\description{
Minimum PSM, Peptide Sequence Match, Count table for
HCC827Ctrl Cell Line.
}
\examples{
{
head(hcc827CtrlPSMCount)
}
}
\references{
Orre et al. 2019 Cell 73, 1-17
}
\keyword{datasets}
