% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc827exon.R
\docType{data}
\name{hcc827exon}
\alias{hcc827exon}
\title{HCC827 Control Exon Cell Line}
\format{
A data frame where 500 exon-centric ensemble identifiers,
corresponding gene symbols, 5 replicated subcellular fractions
and number of unique peptides matched to associated exon.
}
\usage{
hcc827exon
}
\description{
Exon-centric sub data of hcc827 fractionated data.
}
\examples{
{
head(hcc827exon)
}
}
\references{
Orre et al. 2019 Cell 73, 1-17
}
\keyword{datasets}
