% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc827GEF.R
\docType{data}
\name{hcc827GEF}
\alias{hcc827GEF}
\title{Gefitinib treated HCC827 Cell Line}
\format{
A data frame where 10398 protein gene-centric ids and 5
replicated subcellular fractions with duplicates.
}
\usage{
hcc827GEF
}
\description{
HCC827 cell line was treated with Gefitinib which is
EGFR inhibition.
}
\examples{
{
head(hcc827GEF)
}
}
\references{
Orre et al. 2019 Cell 73, 1-17
}
\keyword{datasets}
