% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc827GEFClass.R
\docType{data}
\name{hcc827GEFClass}
\alias{hcc827GEFClass}
\title{Gefitinib treated HCC827 Cell Line Classification}
\format{A data frame where 10398 protein gene-centric ids and corresponding
compartment and neighborhood classification alon with
classification probabilities.}
\usage{
hcc827GEFClass
}
\description{
Gefitinib treated HCC827 cell line classification
contains both neighborhood and compartment level. The data will be
used for the relocalization analysis.
}
\examples{
{
head(hcc827GEFClass)
}
}
\references{
Orre et al. 2019 Cell 73, 1-17
}
\keyword{datasets}
