% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc827GefPSMCount.R
\docType{data}
\name{hcc827GefPSMCount}
\alias{hcc827GefPSMCount}
\title{Minimum PSM Count in HCC827 Gefitinib Cell Line.}
\format{
A data frame where 10398 protein gene-centric ids minimum PSM count.
}
\usage{
hcc827GefPSMCount
}
\description{
Minimum PSM, Peptide Sequence Match, Count table
for HCC827 Gefitinib Cell Line.
}
\examples{
{
head(hcc827GefPSMCount)
}
}
\references{
Orre et al. 2019 Cell 73, 1-17
}
\keyword{datasets}
