% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{loadData}
\alias{loadData}
\title{Load the fractionated proteomics data}
\usage{
loadData(protein.data)
}
\arguments{
\item{protein.data}{data.frame; fractionated proteomics data where
data contains 10 columns of duplicated 5 fractionations and
rownames must be gene-centric protein names}
}
\value{
protein.data.df
}
\description{
Sampled median normalized TMT ratios are checked if
there is any "NA" valeus. If any, the corresponding
row is filtered out. Later, the data is normalized by taking log2.
}
\examples{
{

df <- loadData(SubCellBarCode::hcc827Ctrl[1:10,])
}
}
