% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerProteins.R
\docType{data}
\name{markerProteins}
\alias{markerProteins}
\title{Marker Proteins Source}
\format{A data frame of 3365 proteins as rows and 13 columns headers.}
\usage{
markerProteins
}
\description{
Data for the proteins whose localizations
were well characterized. It also contains color codes for each
compartment and median fractionation profiles for 5 fractions which are
Cyto., Nsol., NucI., Horg., Lorg., with replicates A and B. These
fractionation profiles will be used for the marker protein quality control.
}
\references{
Orre et al. 2019 Cell 73, 1-17
}
\keyword{datasets}
