% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeProbability.R
\name{mergeProbability}
\alias{mergeProbability}
\title{Merge compartment probabilities to neighborhood probabilities}
\usage{
mergeProbability(df)
}
\arguments{
\item{df}{data.frame; all predictions at the neighborhood level and
probablity vectors for each protein}
}
\value{
merged.df
}
\description{
Compartment levels classifications are summed up to
associated neighborhood levels. It is a helper function.
}
\examples{
{

df <- loadData(SubCellBarCode::hcc827Ctrl)

c.prots <- calculateCoveredProtein(rownames(df), markerProteins[,1])

set.seed(7)
c.prots <- sample(c.prots, 365)
cls <- svmClassification(c.prots, df, markerProteins)

all.A <- cls[[1]]$all.prot.pred

all.n.repA <- replacePrediction(all.A, column = "svm.pred.all")

m.all.repA <- mergeProbability(all.n.repA)

}
}
