% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyPlot.R
\name{sankeyPlot}
\alias{sankeyPlot}
\title{Sankey plot for condition-dependent protein relocalization}
\usage{
sankeyPlot(sampleCls1, sampleCls2)
}
\arguments{
\item{sampleCls1}{data.frame; merged classification,
combination of compartment and neighborhood classification.}

\item{sampleCls2}{data.frame; merged classification,
combination of compartment and neighborhood classification.}
}
\value{
label.link.df
}
\description{
Identify candidate condition-dependent relocated
proteins by comparing neighborhood classifications.
}
\examples{
{

exp.cls.df <- SubCellBarCode::hcc827GEFClass

sankeyData <- sankeyPlot(exp.cls.df, exp.cls.df)

}
}
