% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumProbability.R
\name{sumProbability}
\alias{sumProbability}
\title{Sum compartment test data probabilities to
neighborhood probabilities}
\usage{
sumProbability(df)
}
\arguments{
\item{df}{data.frame; test data classifications at the neighborhood
level and probablity vectors for each protein.}
}
\value{
summed.df
}
\description{
Compartment levels classifications on the  test data are
summed up to associated neighborhood levels. It is a helper function.
}
\examples{
{

df <- loadData(SubCellBarCode::hcc827Ctrl)

c.prots <- calculateCoveredProtein(rownames(df), markerProteins[,1])

set.seed(7)
c.prots <- sample(c.prots, 365)
cls <- svmClassification(c.prots, df, markerProteins)

test.A <- cls[[1]]$svm.test.prob.out

rep.obs <- replacePrediction(df = test.A, column = "Observation")

neighborhood.repA <- replacePrediction(df=rep.obs, column = "svm.pred")

sum.repA <- sumProbability(neighborhood.repA)
}
}
