% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCompNeigh.R
\name{mergeCls}
\alias{mergeCls}
\title{Merge compartment and neighborhood classification}
\usage{
mergeCls(compartmentCls, neighborhoodCls)
}
\arguments{
\item{compartmentCls}{data.frame; all predictions, including
unclassified as well, and
probablity vectors for each protein in compartment classification}

\item{neighborhoodCls}{data.frame; all predictions, including
unclassified as well, and
probablity vectors for each protein in compartment classification}
}
\value{
cls.df
}
\description{
Compartment and neighborhood classifications are merged
for the single output.
}
\examples{
{

df <- loadData(SubCellBarCode::hcc827Ctrl)

c.prots <- calculateCoveredProtein(rownames(df), markerProteins[,1])

r.markers <- markerQualityControl(c.prots, df)

cls <- svmClassification(r.markers, df, markerProteins)

test.A <- cls[[1]]$svm.test.prob.out
test.B <- cls[[2]]$svm.test.prob.out

t.c.df <- computeThresholdCompartment(test.A, test.B)

t.n.df <- computeThresholdNeighborhood(test.A, test.B)

all.A <- cls[[1]]$all.prot.pred
all.B <- cls[[2]]$all.prot.pred

c.cls.df <- applyThresholdCompartment(all.A, all.B, t.c.df)

n.cls.df <- applyThresholdNeighborhood(all.A, all.B, t.n.df)

cls.df <- mergeCls(c.cls.df, n.cls.df)
}
}
