% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcode.R
\name{plotBarcode}
\alias{plotBarcode}
\title{Visualize the SubCellBarCode}
\usage{
plotBarcode(sampleClassification, protein, s1PSM)
}
\arguments{
\item{sampleClassification}{data.frame; merged classification,
combination of compartment and neighborhood classification.}

\item{protein}{character; protein gene symbol name}

\item{s1PSM}{data.frame; minimum PSM count table. Row names should
be gene centric protein id.}
}
\value{
proteinPlot
}
\description{
Stacked bar plot are plotted for compartment and
neighborhood level with respect to classification probabilities.
}
\examples{
{

df <- loadData(SubCellBarCode::hcc827Ctrl)

c.prots <- calculateCoveredProtein(rownames(df), markerProteins[,1])

set.seed(7)
c.prots <- sample(c.prots, 365)
cls <- svmClassification(c.prots, df, markerProteins)

test.A <- cls[[1]]$svm.test.prob.out
test.B <- cls[[2]]$svm.test.prob.out

t.c.df <- computeThresholdCompartment(test.A, test.B)

t.n.df <- computeThresholdNeighborhood(test.A, test.B)

all.A <- cls[[1]]$all.prot.pred
all.B <- cls[[2]]$all.prot.pred

c.cls.df <- applyThresholdCompartment(all.A[1:300,],all.B[1:300,],t.c.df)

n.cls.df <- applyThresholdNeighborhood(all.A[1:300,],all.B[1:300,],t.n.df)

cls.df <- mergeCls(c.cls.df, n.cls.df)

proteinPlot <- plotBarcode(cls.df, "ACAA2", hcc827CtrlPSMCount)
}
}
