% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyPlot.R
\name{sankeyPlot}
\alias{sankeyPlot}
\title{Sankey plot for condition-dependent protein relocalization}
\usage{
sankeyPlot(sampleCls1, sampleCls2)
}
\arguments{
\item{sampleCls1}{data.frame; merged classification,
combination of compartment and neighborhood classification.}

\item{sampleCls2}{data.frame; merged classification,
combination of compartment and neighborhood classification.}
}
\value{
sankeyData
}
\description{
Identify candidate condition-dependent relocated
proteins by comparing neighborhood classifications.
}
\examples{
{

df <- loadData(SubCellBarCode::hcc827Ctrl)

c.prots <- calculateCoveredProtein(rownames(df), markerProteins[,1])

r.markers <- markerQualityControl(c.prots, df)

cls <- svmClassification(r.markers, df, markerProteins)

test.A <- cls[[1]]$svm.test.prob.out
test.B <- cls[[2]]$svm.test.prob.out

t.c.df <- computeThresholdCompartment(test.A, test.B)

t.n.df <- computeThresholdNeighborhood(test.A, test.B)

all.A <- cls[[1]]$all.prot.pred
all.B <- cls[[2]]$all.prot.pred

c.cls.df <- applyThresholdCompartment(all.A, all.B, t.c.df)

n.cls.df <- applyThresholdNeighborhood(all.A, all.B, t.n.df)

cls.df  <- mergeCls(c.cls.df, n.cls.df)

sankeyData <- sankeyPlot(cls.df, hcc827GEFClass)

}
}
